Solution to Get nth Bit of a Number

+++ Get nth Bit of a Number@@67882a731b4392d8fc1577ea/code.c
@@ -5,7 +5,7 @@ int main() {
     char str[50];
     scanf("%d %d",&num, &bit);
     sprintf(str, "%d", num);
-    numlen = strlen(num);
+    numlen = (int)strlen(num);
     printf("%d",numlen);
     return 0;
 }
